readable comments to the code
