+++ Backend/models/transporter.js
@@ -1,8 +1,8 @@
 const nodemailer = require('nodemailer');
 
 const transporter = nodemailer.createTransport({
-    host: 'smtp.gmail.com',
-    port: 465,
+    host: 'smtp.gmail.com',// SMTP server for Gmail
+    port: 465, // Port for secure SMTP
     secure: true,
     auth: {
         user: 'agrocom0532@gmail.com', // Your Gmail email or App Password email
@@ -10,4 +10,5 @@ const transporter = nodemailer.createTransport({
     }
 });
 
+// Export the transporter for use in other parts of the application
 module.exports = transporter;
